requirements.txt
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
Faker==38.0.0
